All bugs are fixed. Trying to regulate the GA so it can solve the problem with 20 cities. It has already solved it, at least once, and it gets real close most of the times.

+++ src/br/com/rodrigojunqueira/savanna/core/PrideTest.java
@@ -22,7 +22,7 @@ public class PrideTest {
 	public void nextGenerationCount() {
 		Pride pride = new Pride(new MinusXSquaredFactory());
 		pride.populate(5);
-		int generation = pride.getGeneration();
+		long generation = pride.getGeneration();
 		pride.nextGeneration();
 		assertEquals("After the next generation of the Pride", generation + 1, pride.getGeneration());		
 	}
